[pkg] move src code to module dir
