+++ cdsirius/__init__.py
@@ -1,0 +1,5 @@
+"""
+cdSirius - Compound Discoverer integration with Sirius 6.0 for MS/MS compound annotation
+"""
+
+__version__ = "0.1.0"
