--- node/node_args_mac.json
@@ -1,51 +1,0 @@
-{
-  "CurrentWorkflowID": 140,
-  "ExpectedResponsePath": "/Users/pleeferguson/scratch/node_response.json",
-  "ResultFilePath": "/Users/pleeferguson/Documents/Projects/CleanStreak 2022/CleanStreak_extracts_FISh.cdResult",
-  "NodeParameters": {
-    "Sirius Program Path": "/Applications/sirius-1.localized/sirius.app/Contents/MacOS/sirius",
-    "Save Sirius Result": "False",
-    "Sirius Username": "lee.ferguson@duke.edu",
-    "Sirius Password": "121#Hudson",
-    "Checked Feature Status Handling": "All",
-    "Peak Quality Threshold": "5.5",
-    "Maximum MW": "1500",
-    "Predict Formulas": "True",
-    "Maximum Formula Candidates": "10",
-    "MS1 Mass Accuracy [ppm]": "2",
-    "MS2 Mass Accuracy [ppm]": "5",
-    "Filter by Isotope Pattern": "True",
-    "Enforce Lipid Detection Filtering": "True",
-    "Perform Bottom-Up Formula Search": "True",
-    "De novo Formula Generation Threshold": "400",
-    "Formula Elemental Constraints": "HCNOP[4]F[40]",
-    "Predict Structures": "True",
-    "PubChem as Fallback": "True",
-    "Predict Compound Classes": "True",
-    "Predict de Novo Structures": "False",
-    "De Novo Structure Candidates Limit": "128"
-  },
-  "Version": 1,
-  "Tables": [
-    {
-      "TableName": "Compounds",
-      "DataFile": "/Users/pleeferguson/scratch/ConsolidatedUnknownCompoundItems.txt",
-      "DataFormat": "CSV",
-      "Options": {},
-      "ColumnDescriptions": [
-        {
-          "ColumnName": "Compounds ID",
-          "ID": "ID",
-          "DataType": "Int",
-          "Options": {}
-        },
-        {
-          "ColumnName": "Background",
-          "ID": "",
-          "DataType": "Boolean",
-          "Options": {}
-        }
-      ]
-    }
-  ]
-}
